+++ UrlEncodedFormatter.php
@@ -7,6 +7,7 @@
 
 namespace yii\httpclient;
 
+use Yii;
 use yii\base\Object;
 
 /**
@@ -27,6 +28,10 @@ class UrlEncodedFormatter extends Object implements FormatterInterface
      *    This encoding type is required by OpenID and OAuth protocols.
      */
     public $encodingType = PHP_QUERY_RFC1738;
+    /**
+     * @var string the content charset. If not set, it will use the value of [[\yii\base\Application::charset]].
+     */
+    public $charset;
 
 
     /**
@@ -47,7 +52,9 @@ class UrlEncodedFormatter extends Object implements FormatterInterface
             return $request;
         }
 
-        $request->getHeaders()->set('Content-Type', 'application/x-www-form-urlencoded');
+        $charset = $this->charset === null ? Yii::$app->charset : $this->charset;
+
+        $request->getHeaders()->set('Content-Type', 'application/x-www-form-urlencoded; charset=' . $charset);
         $request->setContent($content);
         return $request;
     }